remove Status enum
