+++ vimgolf/commands/config.py
@@ -1,4 +1,4 @@
-from vimgolf import logger, Status
+from vimgolf import logger, Failure
 from vimgolf.api_key import (
     validate_api_key,
     show_api_key_error,
@@ -13,16 +13,12 @@ def config(api_key=None):
     logger.info('config(...)')
     if api_key is not None and not validate_api_key(api_key):
         show_api_key_error()
-        return Status.FAILURE
-
+        raise Failure()
     if api_key:
         set_api_key(api_key)
-        return Status.SUCCESS
-
-    api_key = get_api_key()
-    if api_key:
-        write(api_key)
     else:
-        show_api_key_help()
-
-    return Status.SUCCESS
+        api_key = get_api_key()
+        if api_key:
+            write(api_key)
+        else:
+            show_api_key_help()
